# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CPI Norte — Sistema Integral de Asistencia
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List all rooms
      parameters:
        - in: query
          name: centerId
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{roomId}:
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete a room (only if no children assigned)
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "409":
          description: Room has children
    patch:
      operationId: updateRoom
      tags: [rooms]
      summary: Update room name or capacity
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomUpdate"
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/summary:
    get:
      operationId: getRoomsSummary
      tags: [rooms]
      summary: Today attendance and vacancy summary per room
      parameters:
        - in: query
          name: centerId
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoomSummary"
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Global attendance and alert stats
      parameters:
        - in: query
          name: centerId
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/alerts:
    get:
      operationId: getAlerts
      tags: [dashboard]
      summary: Children with 2+ consecutive absences
      parameters:
        - in: query
          name: centerId
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
...
components:
  schemas:
...
    StatusResponse:
      type: object
      properties:
        status:
          type: string
      required: [status]
...
    RoomInput:
      type: object
      properties:
        centerId:
          type: integer
        ecoNumber:
          type: integer
        name:
          type: string
        capacity:
          type: integer
      required: [centerId, ecoNumber, name, capacity]
    RoomUpdate:
      type: object
      properties:
        name:
          type: string
        capacity:
          type: integer
    Room:
      type: object
      properties:
        id:
          type: integer
        centerId:
          type: integer
        ecoNumber:
          type: integer
        name:
          type: string
        capacity:
          type: integer
      required: [id, centerId, ecoNumber, name, capacity]
    RoomSummary:
      type: object
      properties:
        id:
          type: integer
        centerId:
          type: integer
        ecoNumber:
          type: integer
        name:
          type: string
        capacity:
          type: integer
        total:
          type: integer
        present:
          type: integer
        absent:
          type: integer
        mercaderia:
          type: integer
        unmarked:
          type: integer
        alerts:
          type: integer
        pct:
          type: integer
      required: [id, centerId, ecoNumber, name, capacity, total, present, absent, mercaderia, unmarked, alerts, pct]
...
    DashboardSummary:
      type: object
      properties:
        totalActive:
          type: integer
        totalPresent:
          type: integer
        totalAbsent:
          type: integer
        totalAlerts:
          type: integer
        totalDischarge:
          type: integer
        totalCapacity:
          type: integer
        pctPresent:
          type: integer
      required: [totalActive, totalPresent, totalAbsent, totalAlerts, totalDischarge, totalCapacity, pctPresent]
    Alert:
      type: object
      properties:
        childId:
          type: integer
        apellido:
          type: string
        nombre:
          type: string
        ecoNumber:
          type: integer
        celular:
          type: ["string", "null"]
        famNombre:
          type: ["string", "null"]
        famApellido:
          type: ["string", "null"]
        consecutiveAbsences:
          type: integer
      required: [childId, apellido, nombre, ecoNumber, consecutiveAbsences]