# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CPI Norte — Sistema Integral de Asistencia
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalActive: number;
  totalPresent: number;
  totalAbsent: number;
  totalAlerts: number;
  totalDischarge: number;
  totalCapacity: number;
  pctPresent: number;
}