# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CPI Norte — Sistema Integral de Asistencia
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: number;
  centerId: number;
  ecoNumber: number;
  name: string;
  capacity: number;
}