# lib/api-zod/src/generated/types/roomSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CPI Norte — Sistema Integral de Asistencia
 * OpenAPI spec version: 0.1.0
 */

export interface RoomSummary {
  id: number;
  centerId: number;
  ecoNumber: number;
  name: string;
  capacity: number;
  total: number;
  present: number;
  absent: number;
  mercaderia: number;
  unmarked: number;
  alerts: number;
  pct: number;
}